:sparkles: created /staff/get endpoint
