+++ backend/src/main/java/com/rota/api/StaffService.java
@@ -50,4 +50,15 @@ public class StaffService {
   public Staff createStaff(Staff newStaff) {
     return staffRepository.save(newStaff);
   }
+
+  /**
+   * Returns the list of all active staff members.
+   *
+   * @return A list of all active staff members.
+   */
+  public List<Staff> getActiveStaff() {
+    return staffRepository.findAll().stream()
+        .filter(staff -> !staff.isInactive())
+        .collect(Collectors.toList());
+  }
 }
